% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSTget.R
\name{DSTget}
\alias{DSTget}
\title{DSTget}
\usage{
DSTget(tablename, language = "en")
}
\arguments{
\item{tablename}{A string giving the name of the table}

\item{language}{A string giving the language used, "en" is suggested}
}
\value{
A table object - which is a list of all metadata
}
\description{
Retrieve table metadata
}
\examples{
bev <- DSTget("BEV3C")
}

