% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixArgEncoding.R
\name{fixArgEncoding}
\alias{fixArgEncoding}
\title{fixArgEncoding}
\usage{
fixArgEncoding(args)
}
\arguments{
\item{args}{The ... arguements from getData}
}
\description{
This function check if any encodings are explicitly given as latin1 and then convert to UTF8
}

