% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSTget.R
\name{summary.DSTget}
\alias{summary.DSTget}
\title{Describe the retrieved table metadata}
\usage{
\method{summary}{DSTget}(table)
}
\arguments{
\item{table}{The R object holding the table information}
}
\value{
Prints description to console
}
\description{
Describe the retrieved table metadata
}
\examples{
myTable <- DSTget("BEV3C")
summary(myTable)
}

